+++ src/components/atomic-crm/payments/index.ts
@@ -1,0 +1,4 @@
+export { AccountPaymentList } from "./AccountPaymentList";
+export { AddPayment } from "./AddPayment";
+export { AccountPaymentEditSheet } from "./AccountPaymentEditSheet";
+export { AccountPaymentInputs } from "./AccountPaymentInputs";